Move `CallGraph` into a new `ipo` module.
